Fix spatial hash float truncation: widen position params to double (end-to-end)
Bug: SpatialHash::Insert/Remove/Update/QueryRadius/GetEntityTargetKey/
GetSubquadrantCoords and the entityPositions map carried positions as
float. At world scale (>1-2 km, per M2.6 double-precision
Transform) float narrowed the authoritative dvec3 position, corrupting
cell keys for far entities. Bug site: old SpatialHash.cpp:134-135
(px/pz float fallback + Insert double-narrowing into
entityPositions), and the float param signatures throughout.

Fix: widen every position parameter and the entityPositions value type
to double (SpatialHash.h:19-31 signatures, .cpp:7-189 all
call sites + Rebucket transform->position.x/z carried as double,
no float narrowing). CELL_SIZE stays a float literal for cell
math; only inputs/keys/stored positions are double. This matches
the M2.6 double-precision direction and keeps hash keys exact
at world scale.

Verification: full build+link in Release+ASan config
(build-asan) — compiles AND links against the widened double
signatures across every caller (not a single-TU ad-hoc compile):
BUILD_EXIT=0, no SpatialHash errors. Canonical suite
ZombieEngineTests.exe: 12 cases / 61 assertions PASS
(TEST_EXIT=0). Verifier persisted:
hermes-verify-spatialhash-asan.sh + /tmp/sh_asan.log.

Note (separate, pre-existing env issue, NOT this change): the
canonical Debug build/ config fails to LINK due to a Catch2
MDd/MD runtime mismatch (LNK2038) — unrelated to this
diff; proven by the clean build-asan link (consistent MD).
Isolated commit per session pattern (push-constant 8809f55,
UAF c7e5000 committed separately).

+++ src/ecs/SpatialHash.cpp
@@ -4,58 +4,59 @@
 
 namespace ecs {
 
-std::pair<int32_t, int32_t> SpatialHash::GetSubquadrantCoords(float x, float z) const {
-    float remainderX = std::fmod(x, CELL_SIZE);
-    float remainderZ = std::fmod(z, CELL_SIZE);
+std::pair<int32_t, int32_t> SpatialHash::GetSubquadrantCoords(double x, double z) const {
+    double remainderX = std::fmod(x, CELL_SIZE);
+    double remainderZ = std::fmod(z, CELL_SIZE);
     if (remainderX < 0) remainderX += CELL_SIZE;
     if (remainderZ < 0) remainderZ += CELL_SIZE;
-    
+
     int32_t subX = static_cast<int32_t>(remainderX / (CELL_SIZE / 4.0f));
     int32_t subZ = static_cast<int32_t>(remainderZ / (CELL_SIZE / 4.0f));
     return { subX, subZ };
 }
 
-uint64_t SpatialHash::GetEntityTargetKey(float x, float z, bool isSubdivided) const {
+uint64_t SpatialHash::GetEntityTargetKey(double x, double z, bool isSubdivided) const {
     uint64_t parentKey = SpatialHashKey(x, z);
     if (!isSubdivided) {
         return parentKey;
     }
-    
+
     auto [subX, subZ] = GetSubquadrantCoords(x, z);
     return ComputeSubdividedCellKey(parentKey, subX, subZ);
 }
 
-void SpatialHash::Insert(entt::entity entity, float x, float z, entt::registry* registry) {
+void SpatialHash::Insert(entt::entity entity, double x, double z, entt::registry* registry) {
     uint64_t parentKey = SpatialHashKey(x, z);
     auto& header = headers[parentKey];
-    
+
     header.entityCount++;
     entityPositions[entity] = {x, z};
-    
+
     // Check if we need to subdivide
     if (!header.isCellSubdivided && header.entityCount > SUBDIVISION_THRESHOLD) {
         header.isCellSubdivided = true;
-        
+
         // Re-bucket existing entities
         std::vector<entt::entity> tempEntities = std::move(buckets[parentKey]);
         buckets.erase(parentKey);
-        
+
         for (entt::entity ent : tempEntities) {
-            float px = 0.0f, pz = 0.0f;
+            double px = 0.0, pz = 0.0;
             bool foundPos = false;
-            
+
             // [M1-EXT-08] Recompute leaf quadrant using the entity's actual position if registry is provided
             if (registry && registry->valid(ent)) {
                 auto* transform = registry->try_get<Transform>(ent);
                 if (transform) {
+                    // Carry the authoritative dvec3 position as double (no float narrowing).
                     px = transform->position.x;
                     pz = transform->position.z;
                     foundPos = true;
                     // Update cache to match actual position
-                    entityPositions[ent] = {px, pz}; 
+                    entityPositions[ent] = {px, pz};
                 }
             }
-            
+
             // Fallback to cached position if no registry or transform
             if (!foundPos) {
                 auto posIt = entityPositions.find(ent);
@@ -69,17 +70,17 @@ void SpatialHash::Insert(entt::entity entity, float x, float z, entt::registry* 
             if (foundPos) {
                 uint64_t targetKey = GetEntityTargetKey(px, pz, true);
                 buckets[targetKey].push_back(ent);
-                
+
                 auto [subX, subZ] = GetSubquadrantCoords(px, pz);
                 uint32_t leafOffset = ((static_cast<uint32_t>(subX) & 0x03) << 2) | (static_cast<uint32_t>(subZ) & 0x03);
                 header.subQuadrantMask |= (1 << leafOffset);
             }
         }
     }
-    
+
     uint64_t targetKey = GetEntityTargetKey(x, z, header.isCellSubdivided);
     buckets[targetKey].push_back(entity);
-    
+
     if (header.isCellSubdivided) {
         auto [subX, subZ] = GetSubquadrantCoords(x, z);
         uint32_t leafOffset = ((static_cast<uint32_t>(subX) & 0x03) << 2) | (static_cast<uint32_t>(subZ) & 0x03);
@@ -87,17 +88,17 @@ void SpatialHash::Insert(entt::entity entity, float x, float z, entt::registry* 
     }
 }
 
-void SpatialHash::Remove(entt::entity entity, float x, float z) {
+void SpatialHash::Remove(entt::entity entity, double x, double z) {
     uint64_t parentKey = SpatialHashKey(x, z);
     auto it = headers.find(parentKey);
     if (it == headers.end()) return;
-    
+
     auto& header = it->second;
     header.entityCount--;
-    
+
     uint64_t targetKey = GetEntityTargetKey(x, z, header.isCellSubdivided);
     auto& bucket = buckets[targetKey];
-    
+
     auto entIt = std::find(bucket.begin(), bucket.end(), entity);
     if (entIt != bucket.end()) {
         bucket.erase(entIt);
@@ -105,7 +106,7 @@ void SpatialHash::Remove(entt::entity entity, float x, float z) {
     entityPositions.erase(entity);
 }
 
-void SpatialHash::Update(entt::entity entity, float oldX, float oldZ, float newX, float newZ, entt::registry* registry) {
+void SpatialHash::Update(entt::entity entity, double oldX, double oldZ, double newX, double newZ, entt::registry* registry) {
     Remove(entity, oldX, oldZ);
     Insert(entity, newX, newZ, registry);
 }
@@ -117,9 +118,9 @@ void SpatialHash::Clear() {
 }
 
 void SpatialHash::Rebucket(entt::registry& registry) {
-    std::vector<std::tuple<entt::entity, float, float, float, float>> toUpdate;
-    std::vector<std::tuple<entt::entity, float, float>> toRemove;
-    
+    std::vector<std::tuple<entt::entity, double, double, double, double>> toUpdate;
+    std::vector<std::tuple<entt::entity, double, double>> toRemove;
+
     for (const auto& [entity, pos] : entityPositions) {
         if (!registry.valid(entity)) {
             toRemove.push_back({entity, pos.first, pos.second});
@@ -130,14 +131,15 @@ void SpatialHash::Rebucket(entt::registry& registry) {
             toRemove.push_back({entity, pos.first, pos.second});
             continue;
         }
-        
-        float newX = transform->position.x;
-        float newZ = transform->position.z;
+
+        // Carry the authoritative dvec3 position as double (no float narrowing).
+        double newX = transform->position.x;
+        double newZ = transform->position.z;
         if (newX != pos.first || newZ != pos.second) {
             toUpdate.push_back({entity, pos.first, pos.second, newX, newZ});
         }
     }
-    
+
     for (const auto& rm : toRemove) {
         Remove(std::get<0>(rm), std::get<1>(rm), std::get<2>(rm));
     }
@@ -149,10 +151,10 @@ void SpatialHash::Rebucket(entt::registry& registry) {
 std::vector<entt::entity> SpatialHash::QueryCell(int32_t cx, int32_t cz) const {
     uint64_t parentKey = (static_cast<uint64_t>(cx) << 32) | (static_cast<uint64_t>(cz) & 0xFFFFFFFFull);
     std::vector<entt::entity> result;
-    
+
     auto it = headers.find(parentKey);
     if (it == headers.end()) return result;
-    
+
     if (!it->second.isCellSubdivided) {
         auto bucketIt = buckets.find(parentKey);
         if (bucketIt != buckets.end()) {
@@ -170,25 +172,25 @@ std::vector<entt::entity> SpatialHash::QueryCell(int32_t cx, int32_t cz) const {
             }
         }
     }
-    
+
     return result;
 }
 
-std::vector<entt::entity> SpatialHash::QueryRadius(float x, float z, float radius) const {
+std::vector<entt::entity> SpatialHash::QueryRadius(double x, double z, double radius) const {
     std::vector<entt::entity> result;
-    
+
     int32_t minCx = static_cast<int32_t>(std::floor((x - radius) / CELL_SIZE));
     int32_t maxCx = static_cast<int32_t>(std::floor((x + radius) / CELL_SIZE));
     int32_t minCz = static_cast<int32_t>(std::floor((z - radius) / CELL_SIZE));
     int32_t maxCz = static_cast<int32_t>(std::floor((z + radius) / CELL_SIZE));
-    
+
     for (int32_t cx = minCx; cx <= maxCx; ++cx) {
         for (int32_t cz = minCz; cz <= maxCz; ++cz) {
             auto cellResult = QueryCell(cx, cz);
             result.insert(result.end(), cellResult.begin(), cellResult.end());
         }
     }
-    
+
     // NOTE: This returns all entities in overlapping cells.
     // Consumer code should do exact distance filtering.
     return result;
